Move MBC modules and ModbusDataType into src/MbcInterface/
All MBC-specific source files (mbcfileimporter, mbcregisterdata,
mbcregistermodel, mbcregisterfilter, mbcupdatemodel, mbcheader) and
ModbusDataType are consolidated under src/MbcInterface/ to make the
protocol boundary explicit.

Unused methods removed from ModbusDataType: isUnsigned, isFloat,
description, convertSettings, convertString, plus the corresponding
bUnsigned and bFloat fields from the TypeSettings struct.

Tests moved to tests/MbcInterface/ with CMakeLists updated accordingly.
