+++ src/dialogs/importmbcdialog.cpp
@@ -1,14 +1,14 @@
 #include "importmbcdialog.h"
 
 #include "customwidgets/actionbuttondelegate.h"
-#include "dialogs/mbcheader.h"
 #include "dialogs/ui_importmbcdialog.h"
-#include "importexport/mbcfileimporter.h"
+#include "MbcInterface/mbcfileimporter.h"
+#include "MbcInterface/mbcheader.h"
+#include "MbcInterface/mbcregisterfilter.h"
+#include "MbcInterface/mbcregistermodel.h"
+#include "MbcInterface/mbcupdatemodel.h"
 #include "models/graphdatamodel.h"
 #include "models/guimodel.h"
-#include "models/mbcregisterfilter.h"
-#include "models/mbcregistermodel.h"
-#include "models/mbcupdatemodel.h"
 #include "util/fileselectionhelper.h"
 #include "util/util.h"
