+++ tests/CMakeLists.txt
@@ -52,6 +52,7 @@ add_subdirectory(customwidgets)
 add_subdirectory(datahandling)
 add_subdirectory(dialogs)
 add_subdirectory(importexport)
+add_subdirectory(MbcInterface)
 add_subdirectory(models)
 add_subdirectory(ProtocolAdapter)
 add_subdirectory(util)
