+++ tests/importexport/CMakeLists.txt
@@ -1,6 +1,4 @@
 
-add_xtest_mock(tst_mbcfileimporter ${CMAKE_CURRENT_SOURCE_DIR}/mbctestdata.cpp)
-add_xtest_mock(tst_mbcregisterfilter)
 add_xtest(tst_datafileparser ${CMAKE_CURRENT_SOURCE_DIR}/csvdata.cpp)
 add_xtest_mock(tst_presethandler)
 add_xtest(tst_presetparser ${CMAKE_CURRENT_SOURCE_DIR}/presetfiletestdata.cpp)
